majorly refactors History, Move, HistoryItem, Variation, Fullmove, MoveColView, MoveListColView and VariationColView
turns all pointers into private members with get/set methods.  changes the way col view moves are handled with
regards to setting the Fullmove's fullmove number but doesn't fix it properly.

this is a checkpoint commit before doing drastic changes to Variation - updatePointers will be removed as most of it
is useless (halfmove, fullmove, colour will never be changed - only nextItem/prevItem need to be updated so it can be
done in each history manipulation function)