changes User and Application so that a new User is created on each connection
this fixes the problem where the user connects again, and there is a User object for
them in their current Client's session, and that User has a reference to whichever client
was passed to it when the user first connected, which probably isn't still connected.

Now a new user is created each time.  encapsulation is better as the User will handle
the client disconnecting and the job of looking in the client's session to see if the
user is logged in.