adds keepalive logic and resets interval to 1000. this is necessary for the client.
really it should be called the "server heartbeat" because it's to let the client know
that they are still connected to the internet and the server is still online.  chrome
doesn't know until minutes later that the connection is lost even if the network interface
goes down, so this is the only solution for now.