separates NewGame into two different events in client Application - GamesReceived and GameReady
this is a dirty-ish way of distinguishing between games that are received because the client is
involved in them in some way and called /request/games, and games that are received because they
have just started and the user is playing in them.  In the first case all that needs to happen is
links to the games getting added to the nav bar, but in the second one the game needs to be gone
to as well.  The only real problem is the names of the events not really being sensible, and maybe
the urls "/games" (first case) and "/game" (second case) don't mean enough either.