adds challenge cancel functionality to CreateChallengeForm
the button currently goes transparent and changes to 'Cancel', but it needs
animating with a draining-out effect the length of the challenge timeout.

this turns out not to work very well with css animations; the browser doesn't
seem to smooth out the transitions between the keyframes so you would need
hundreds to make the animation smooth.  should be easy enough to come up
with a js solution.