removes AboutPage and footer from Lightsquare
this stuff should be in a separate site; the about page is too
unrelated to the actual chess app so it clutters up the Lightsquare
widget with all the requires etc.

the app should be a subdomain of the main site, e.g. play.lightsquare
and stuff should be moved up into the main site where appropriate.