adds _createGame to User and handles Rematch in _addGame so that getGame() resolves straight away when other handlers of Rematch get the game out of the User.
This relies on event handler execution order but it isn't critical, it just avoids
sending /spectate/game and pushing another /game/id onto the url history.  It's also reliable
because the User is the one creating the Game, so it knows no handlers have been added yet.

An alternative solution would be for Lightsquare to use the Game available in its own GamePage.Rematch
handler, instead of just navigate()-ing to the new /game/id and letting the Router handler sort
it out.