adds game restoration modal dialog framework to HomePage.  game restoration not implemented yet.
the dark overlay and dialog container are re-usable - other dialogs can
be added as children of the home_page_modal_dialog_container div and will
work as dialogs.  they will be in a 0 height div at the top of the home
page, so just need a top margin and 'auto' right/left margins for centering.

showing a dialog/hiding all dialogs is already implemented.  it is currently
done as a hash of dialog names to booleans, but could probably be
refactored as a single field containing either the name of the currently
showing dialog or null.