implements new game/cancel on GamePage - uses same mechanism and animation system as SeekForm
there is duplication in the button style and the expiry time calculations,
but factoring it out seems too far for now.

the seek form on the homepage updates as appropriate when the user creates
a new seek from the gamepage.  if the user creates a new seek from the seek
form with the same time options as a gamepage, the game page will update as
though the new game button had been pressed.