continues refactoring widgets
moves more code and widgets from HomePage into holding area (snippets
and widgets/)

passes Play a prefixed Router - the Play page will be under /play/...
but it gets a Router with the prefix set to /play so it still only
has to deal with /game/123 etc (but the actual urls will be /play/game/123).
the currentPath on a prefixed Router where the url doesn't have the
prefix will be null, and it won't send UrlChanged events or bother
trying to match the route when the url changes to another one that
doesn't match the prefix.

cleans up HomePage.  splits the main nav bar into nav and user menu.