allows vertical scrollbar on GamePage main container
this solves the problem of wanting scrolling as well as automatic
full-height to make things stretch.  scrolling must be on inner things
as opposed to the whole page, which works well with the layout anyway.

also sets position: relative on the scrolling element so that the board
scrolls with it.