makes disconnected message less intrusive
this means that the UI is still there when the server isn't connected, so
the user can click around and it will be unresponsive.  the advantage is
that they aren't jarringly switched to the big white disconnection page, and
if they're in a game they can still be looking at the position while the
app reconnects.