refactors games management code into a separate module (_Games)
Not sure how much I like this.  Most of the reasoning behind it is
keeping the length of the User module down.  There is a case for
keeping the related functionality separate and together as well,
and this is definitely the best way for the games backup code, but
the games list and seek functions seem to belong in the User more.